fix: Refactor
